+++ data/src/main/java/com/dmitron/data/local/mappers/facade/DatabaseCityWeatherMapperFacade.kt
@@ -9,10 +9,10 @@ class DatabaseCityWeatherMapperFacade(
     val mapDatabaseCityWeather: (DatabaseCityWeather) -> CityWeather,
     val mapCityToDatabase: (City) -> DatabaseCity,
     val mapWeatherToDatabase: (Weather, cityId: Long) -> DatabaseWeather,
-    val mapDayToDatabase: (Weather.Day, weatherId: Int, cityId: Long) -> DatabaseDay,
+    val mapDayToDatabase: (Weather.Day, weatherId: Long, cityId: Long) -> DatabaseDay,
     val mapHourToDatabase: (
         Weather.Day.HourlyWeather, dayId: Long, cityId: Long,
         parentDayOfWeek: Int,
-        parentWeatherId: Int
+        parentWeatherId: Long
     ) -> DatabaseHourlyWeather,
 )
